Move hardware interfacing to subfolder.
