+++ hardware/SPEC.py
