+++ hardware/SpecClient/.gitignore
