+++ hardware/SpecClient/SpecArray.py
