+++ hardware/SpecClient/SpecClientError.py
