+++ hardware/SpecClient/SpecCommand.py
