+++ hardware/SpecClient/SpecConnectionsManager.py
