+++ hardware/SpecClient/SpecMessage.py
