+++ hardware/SpecClient/SpecReply.py
