+++ hardware/SpecClient/SpecServer.py
