+++ hardware/SpecClient/SpecVariable.py
