+++ hardware/SpecClient/clientutils.mac
