+++ hardware/firmware.bin
